Postfix listen on all the interfaces
When you are running the container in a IPv6 enabled
OS but no IPv6 is available in the container, Postfix
will crash with a message like:
"fatal: parameter inet_interfaces: no local interface found for ::1"

Listening an all interfaces is the default for Postfix,
RedHat/CentOS has decided to change this. Putting it back to the default
should not cause issues as the container should not be accessible from
everywhere.